This commit was manufactured by cvs2git to create tag 'HLTriggerOffline- Tau-V02-03-03'.
Sprout from master 2008-07-18 18:59:46 UTC Michail Bachtis <sha1-eaf6ea55826e5289c86058cfc492a286c076afdf@cern.ch> 'Added e/mu'
Cherrypick from master 2008-06-10 21:53:36 UTC Chi Nhan Nguyen <sha1-d7c3810b3b8595346f468e2d6413da596572dbe0@cern.ch> 'Changed label names.':
    data/HLTTauRefInfo.cfi
    interface/HLTTauAnalyzer.h
    interface/HLTTauMcInfo.h
    interface/HLTTauRefInfo.h
    interface/HLTTauValidation.h
    interface/L25TauValidation.h
    interface/L2TauValidation.h
    src/HLTTauAnalyzer.cc
    src/HLTTauL25Validation.cc
    src/HLTTauMcInfo.cc
    src/HLTTauRefInfo.cc
    src/HLTTauValidation.cc
    src/L25TauValidation.cc
    src/L2TauValidation.cc
    src/SealModule.cc
    test/FullHLTValidation_HCh200.cfg
    test/FullHLTValidation_ZTT.cfg
    test/HLTValPlots.cxx
    test/RunL1Analyzer.cfg
    test/RunL2Validation.cfg
    test/drawEfficiencies.C~
    test/drawEfficiency.C~

+++ src/HLTTauMcInfo.cc
@@ -4,6 +4,7 @@
 #include "DataFormats/TauReco/interface/PFTau.h"
 #include "DataFormats/TauReco/interface/PFTauDiscriminatorByIsolation.h"
 #include "SimDataFormats/HepMCProduct/interface/HepMCProduct.h"
+#include "DataFormats/HepMCCandidate/interface/GenParticleCandidate.h"
 
 using namespace edm;
 using namespace reco;
@@ -110,7 +111,7 @@ void HLTTauMcInfo::produce(edm::Event& iEvent, const edm::EventSetup& iES)
       Handle<PFTauDiscriminatorByIsolation> thePFTauDiscriminatorByIsolation;
       iEvent.getByLabel(pfTauDiscriminatorProd_,thePFTauDiscriminatorByIsolation);
 
-      for(size_t it =0; it<product_Jets_tmp.size();it++)
+      for(int it =0; it<product_Jets_tmp.size();it++)
 	{ 
 	  for (PFTauCollection::size_type iPFTau=0;iPFTau<thePFTauHandle->size();iPFTau++) {
 	    PFTauRef thePFTau(thePFTauHandle,iPFTau);
@@ -124,7 +125,7 @@ void HLTTauMcInfo::produce(edm::Event& iEvent, const edm::EventSetup& iES)
 	}
       
     }else{
-      for(size_t it =0; it<product_Jets_tmp.size();it++)
+      for(int it =0; it<product_Jets_tmp.size();it++)
 	{ 
 	  product_Jets->push_back(product_Jets_tmp[it]);
 	}
